Day 9

+++ AdventOfCode/Day5/Part1.cs
@@ -5,6 +5,6 @@ namespace AdventOfCode.Day5
 {
     public static class Part1
     {
-        public static int Solve () => new TestComputer (Data.DiagnosticProgram).Execute ();
+        public static int Solve () => (int) new TestComputer (Data.DiagnosticProgram).Execute ();
     }
 }
