+++ AdventOfCode/Day5/Part2.cs
@@ -2,6 +2,6 @@ namespace AdventOfCode.Day5
 {
     public static class Part2
     {
-        public static int Solve () => new TestComputer (Data.DiagnosticProgram).Execute ();
+        public static int Solve () => (int) new TestComputer (Data.DiagnosticProgram).Execute ();
     }
 }
