+++ AdventOfCode/Day7/Part1.cs
@@ -7,9 +7,9 @@ namespace AdventOfCode.Day7
 {
     public static class Part1
     {
-        public static int Solve () => GetMaxPermutationValue (new HashSet <byte> {0, 1, 2, 3, 4});
+        public static int Solve () => (int) GetMaxPermutationValue (new HashSet <byte> {0, 1, 2, 3, 4});
 
-        private static int GetMaxPermutationValue (HashSet <byte> numbers) => Permute (numbers).Select (EvaluateOutputSignal).Max ();
+        private static long GetMaxPermutationValue (HashSet <byte> numbers) => Permute (numbers).Select (EvaluateOutputSignal).Max ();
 
         public static HashSet <List <byte>> Permute (HashSet <byte> input)
             => input.Count > 1
@@ -24,18 +24,18 @@ namespace AdventOfCode.Day7
                 }).ToHashSet ()
                 : new HashSet <List <byte>> {input.ToList ()};
 
-        private static int EvaluateOutputSignal (List <byte> permutation)
+        private static long EvaluateOutputSignal (List <byte> permutation)
         {
             var writeStack = new Stack <byte> (permutation);
-            var readQueue  = new Queue <int> (1);
+            var readQueue  = new Queue<long>(1);
             readQueue.Enqueue (0);
 
             while (writeStack.Count > 0)
             {
                 var step = 0;
 
-                var reader = new Action <int> (v => readQueue.Enqueue (v));
-                var writer = new Func <int> (() => step++ == 0 ? writeStack.Pop () : readQueue.Dequeue ());
+                var reader = new Action<long>(v => readQueue.Enqueue (v));
+                var writer = new Func <long> (() => step++ == 0 ? writeStack.Pop () : readQueue.Dequeue ());
 
                 var computer = new PipedComputer (Data.Code, reader, writer);
                 computer.Execute ();
